receiver-function/requirements.txt
google-cloud-pubsub==2.18.4
functions-framework==3.5.0
flask==3.0.0